Function chakracore_sys::JsDiagGetObjectFromHandle

pub unsafe extern "system" fn JsDiagGetObjectFromHandle(
    objectHandle: c_uint, 
    handleObject: *mut JsValueRef
) -> JsErrorCode

Gets the object corresponding to handle.
Handle of object. Object corresponding to the handle. { "scriptId" : 24, "line" : 1, "column" : 63, "name" : "foo", "type" : "function", "handle" : 2 } The code JsNoError if the operation succeeded, a failure code otherwise. The current runtime should be in debug state. This API can only be called when runtime is at a break.