Function chakracore_sys::JsDiagGetScripts

pub unsafe extern "system" fn JsDiagGetScripts(
    scriptsArray: *mut JsValueRef
) -> JsErrorCode

Gets list of scripts.
Array of script objects. [{ "scriptId" : 2, "fileName" : "c:\Test\Test.js", "lineCount" : 4, "sourceLength" : 111 }, { "scriptId" : 3, "parentScriptId" : 2, "scriptType" : "eval code", "lineCount" : 1, "sourceLength" : 12 }] The code JsNoError if the operation succeeded, a failure code otherwise. The current runtime should be in debug state. This API can be called when runtime is at a break or running.