Function chakracore_sys::JsDiagGetStackProperties

pub unsafe extern "system" fn JsDiagGetStackProperties(
    stackFrameIndex: c_uint, 
    properties: *mut JsValueRef
) -> JsErrorCode

Gets the list of properties corresponding to the frame.
Index of stack frame from JsDiagGetStackTrace. Object of properties array (properties, scopes and globals). propertyAttributes is a bit mask of NONE = 0x1, HAVE_CHILDRENS = 0x2, READ_ONLY_VALUE = 0x4, { "thisObject": { "name": "this", "type" : "object", "className" : "Object", "display" : "{...}", "propertyAttributes" : 1, "handle" : 306 }, "exception" : { "name" : "{exception}", "type" : "object", "display" : "'a' is undefined", "className" : "Error", "propertyAttributes" : 1, "handle" : 307 } "arguments" : { "name" : "arguments", "type" : "object", "display" : "{...}", "className" : "Object", "propertyAttributes" : 1, "handle" : 190 }, "returnValue" : { "name" : "[Return value]", "type" : "undefined", "propertyAttributes" : 0, "handle" : 192 }, "functionCallsReturn" : [{ "name" : "[foo1 returned]", "type" : "number", "value" : 1, "propertyAttributes" : 2, "handle" : 191 } ], "locals" : [], "scopes" : [{ "index" : 0, "handle" : 193 } ], "globals" : { "handle" : 194 } } The code JsNoError if the operation succeeded, a failure code otherwise. The current runtime should be in debug state. This API can only be called when runtime is at a break.