Function chakracore_sys::JsHasException

pub unsafe extern "system" fn JsHasException(
    hasException: *mut bool
) -> JsErrorCode

Determines whether the runtime of the current context is in an exception state.
If a call into the runtime results in an exception (either as the result of running a script or due to something like a conversion failure), the runtime is placed into an "exception state." All calls into any context created by the runtime (except for the exception APIs) will fail with JsErrorInExceptionState until the exception is cleared. If the runtime of the current context is in the exception state when a callback returns into the engine, the engine will automatically rethrow the exception. Requires an active script context. Whether the runtime of the current context is in the exception state. The code JsNoError if the operation succeeded, a failure code otherwise.