Function chakracore_sys::JsNumberToInt

pub unsafe extern "system" fn JsNumberToInt(
    value: JsValueRef, 
    intValue: *mut c_int
) -> JsErrorCode

Retrieves the int value of a number value.
This function retrieves the value of a number value and converts to an int value. It will fail with JsErrorInvalidArgument if the type of the value is not number. The number value to convert to an int value. The int value. The code JsNoError if the operation succeeded, a failure code otherwise.