Function chakracore_sys::JsParseModuleSource

pub unsafe extern "system" fn JsParseModuleSource(
    requestModule: JsModuleRecord, 
    sourceContext: JsSourceContext, 
    script: *mut BYTE, 
    scriptLength: c_uint, 
    sourceFlag: JsParseModuleSourceFlags, 
    exceptionValueRef: *mut JsValueRef
) -> JsErrorCode

Parse the module source
This is basically ParseModule operation in ES6 spec. It is slightly different in that the ModuleRecord was initialized earlier, and passed in as an argument. The ModuleRecord that holds the parse tree of the source code. A cookie identifying the script that can be used by debuggable script contexts. The source script to be parsed, but not executed in this code. The source length of sourceText. The input might contain embedded null. The type of the source code passed in. It could be UNICODE or utf8 at this time. The error object if there is parse error. The code JsNoError if the operation succeeded, a failure code otherwise.