Function chakracore_sys::JsTTDCreateRecordRuntime

pub unsafe extern "system" fn JsTTDCreateRecordRuntime(
    attributes: JsRuntimeAttributes, 
    snapInterval: usize, 
    snapHistoryLength: usize, 
    openResourceStream: TTDOpenResourceStreamCallback, 
    writeBytesToStream: JsTTDWriteBytesToStreamCallback, 
    flushAndCloseStream: JsTTDFlushAndCloseStreamCallback, 
    threadService: JsThreadServiceCallback, 
    runtime: *mut JsRuntimeHandle
) -> JsErrorCode

TTD API -- may change in future versions: Creates a new runtime in Record Mode.
The attributes of the runtime to be created. The interval to wait between snapshots (measured in millis). The amount of history to maintain before discarding -- measured in number of snapshots and controls how far back in time a trace can be reversed. The TTDOpenResourceStreamCallback function for generating a JsTTDStreamHandle to read/write serialized data. The JsTTDWriteBytesToStreamCallback function for writing bytes to a JsTTDStreamHandle. The JsTTDFlushAndCloseStreamCallback function for flushing and closing a JsTTDStreamHandle as needed. The thread service for the runtime. Can be null. The runtime created. See JsCreateRuntime for additional information. The code JsNoError if the operation succeeded, a failure code otherwise.