Function chakracore_sys::JsTTDReplayExecution

pub unsafe extern "system" fn JsTTDReplayExecution(
    moveMode: *mut JsTTDMoveMode, 
    rootEventTime: *mut i64
) -> JsErrorCode

TTD API -- may change in future versions: Execute from the current point in the log to the end returning the error code.
Additional flags for controling how the move is done. The event time that we should move to next or notification (-1) that replay has ended. If the debugger requested an abort the code is JsNoError -- rootEventTime is the target event time we need to move to and re - execute from. If we aborted at the end of the replay log the code is JsNoError -- rootEventTime is -1. If there was an unhandled script exception the code is JsErrorCategoryScript.