Type Definition chakracore_sys::TTDOpenResourceStreamCallback

type TTDOpenResourceStreamCallback = Option<unsafe extern "system" fn(_: usize, _: *const c_char, _: usize, _: *const c_char, _: bool, _: bool) -> JsTTDStreamHandle>;

TTD API -- may change in future versions: Construct a JsTTDStreamHandle that will be used to read/write the event log portion of the TTD data based on the uri provided by JsTTDInitializeUriCallback.
Exactly one of read or write will be set to true. The length of the uri array that the host passed in for storing log info. The URI that the host passed in for storing log info. The length of the ascii name array that the host passed in for storing log info. An optional ascii string giving a unique name to the resource that the JsTTDStreamHandle will be created for. If the handle should be opened for reading. If the handle should be opened for writing. A JsTTDStreamHandle opened in read/write mode as specified.